new understanding, lots of refactoring
Understanding how to hook in to Jupyter's signals in React via
UseSignal allows a lot of untangling of logic. There's now a pretty
clear Phosphor widget setup layer that handles that and only that. And
the rest of the logic is now much more compartmentalized in only the
React components that care about certain state conditions.