cell inspection and client connection more generic
The logic for child components needing to know when a cell changes (or
the content in it changes) is going to be the same. Plus it's a bit
weird to set up. JupyterCellInspector generalizes this setup, allowing
a functional child component to pass the active cell down to further
interested children.

ADIClient uses the settings context to listen for updates to the
settings and set up the apollo client connection for any children.