remove proof of concept code
Early on in the plugin development, I tested storing and retrieving
metadata by incrementing a counter every time the selection changed.
This means notebooks would show up as needing to be saved when you only
changed selection of cells. We're now actually using the metadata to
store uuid values, so even as an example, the proof of concept code has
outlived its usefulness and is just becoming annoying. So long!