shows how we can manipulate the entire notebook
This will be a throwaway branch, but since some of the API
documentation for JupyterLabe still seems a bit sparse from our point
on the learning curve, I want to keep a record of my mini-investigation.

The work here is related to ADATAI-157, where we were originally
looking at adding functionality to create computed datasets with
reusable transformations and input datasets via JupyterLabe, similar
to how we allow the reusable transformations themselves to be created.
There is a strong use case for this:

A data scientiest working in a JupyterLab environment is developing a
reusable transformation along with one or more computed datasets that
use that transformation. It would be nice to recognize that a method in
the notebook that they're using is actually a transformation, which
could then be used to infer that in another cell, where they're using
that method with some parameters to compute a test dataset that
they're actually showing how they intend to hook up the transformation.
If we know that an input parameter is a dataset obtained from ADI, we
could set up the entire set of parameters to actually define the
computation in ADI. However, that's a lot of things to have to infer,
probably stretching the limits of some of the current solutions for
inferring basic transformation info.

That's why this is being abandonned for now. It's a really interesting
use case, but it's a bit of a niche use case. And it's hard to justify
when the basic use case hasn't been addressed yet.

However, when we get back to it, one thing we'll need to do is be able
to know, at the notebook level, all of the transformations that may
exist in it, so that we can recognize that a particular method being
used also actually represents an ADI transformation. This code shows
how we can scan through all cells and collect some piece of metadata
from them. Note that we'll probably also have to add a little more
metadata than the uuid (like also the last known method name that
uuid was associated with) to make something like this work.